requirements.txt
polars
numpy
python-dotenv
questdb
psycopg2-binary
pandas
matplotlib
mplfinance
